Added 6 entries to main to take care of the "possessive" issue:
A friend of mime told me that when a few entries in Main are followed by apostrophe s "'s", they are read a bit oddly as the 's part gets separated from the words preceding it. So I'm adding 6 entries to Main to take care of the ones which are likely to be used with 's.
Centcom's	`[.1sEnt.2kamz]
CENTCOM's	`[.1sEnt.2kamz]
CSUN's	`[.1si.2sHnz]
Macron's	`[.2mA.1kranz]
TESOL's	`[.1tE.2sclz]
TEFL's	`[.1tE.0fxlz]