Added "Parton", capitalized, to Main:
Parton = `[.1par.0?N]
This was added to take care of the "Dolly Parton" pronunciation issue. We have a "parton entry in Root which alters the pronunciation to `[.1par.2tan] to take care of its pronunciation taken from the following: parton. noun.  Any of the point particles proposed as a constituent of hadrons, now known as quarks.
Are you satisfied with these modifications?