ENURoot: removed couple of duplicates, with comments:
compand: we already have it.
bios: we have it in ENUmain, all caps. The problem with this one in ENURoot is that bios, lowercase or capitalized, is the plural of bio -- the short form of biography. So if this is kept in Root, the plural of bio -- lowercase/capitalized -- will be mispronounced. So, folks, do you want to have "bios" in Root retained?