NPR radio stations added, reorganized stations
I've added all NPR radio stations up to connecticut, so far. I also grouped radio stations in main in one location. Right now they are at the bottom of the file and that won't stay the same but they're all together now. Goal is to add all mispronounced NPR stations (only adding NPR because that's what was in here to start with). Let me know if I'm overdoing it.