removed 1, fixed 1
Thanks for all the tips. I keep messing up the x / X schwa sounds. Beat me with a stick from time to time and I will learn--sorry about that.

Changes are as follows: What I thought was gomphiosis is actually not a word. Gomphosis is the actual word and it as removed.
I fixed chaetiferois as directed. Again, thanks for the guidance.
For future submissions, I'll continue to put the word and pronunciation in the commit for reference and in case I've screwed something up--most likely I'd only submit two or three words at a time.