removed one word and added two others, please read description
tweeked advertisement again. It is more true to form if we leave it the way Eloquence was created :).
Added Blytheville--yes there's a town that pronounces it this way.
Added remembrall from Harry Potter
Also, I added the following:
Epicaricacy	`[1Ep.0xk.0xr.1Ik.0kx.2si]
This is a real word and I'm not sure if I've got the syllables correct. Please verify.