Very serious issue affecting words with special characters fixed:
This issue which affects words with non-English letters popped up after the release of V2023.07, so last month's release isn't affected. It might be owing to the fact that now many users contribute to the project and special characters might not be supported on some of the machines, resulting in the appearance of question marks instead of those characters. From now on I'll make sure to check those before each commit.